Add ElfDynamicSection and accompanying tests
Implement `ElfDynamicSection` as a subclass of `ElfSection` and add it
to the list of sections as we're parsing the section table. Technically
the dynamic section entries should have a union of value or pointer but
for simplicity we just call everything "Value".

For entries that are offsets into a string table we parse the string
out as an additional pass after parsing all sections.